scx_utils: Move perf event code to perf crate
scx_cosmos implements specific perf event spec parsing and PMU setup
code that could be used by other scx schedulers. Move it into the
existing scx_utils::perf crate to prevent code duplication.

The shared code uses scx_utils own perf bindings and perf_event_open(),
instead of the external perf-event-open-sys crate, which is dropped from
scx_cosmos.

setup_perf_events() takes the PMU map directly (&impl MapCore) rather
than a crate-specific BpfSkel, so it can live in the shared crate.
