+++ scheds/rust/scx_cosmos/Cargo.toml
@@ -15,7 +15,6 @@ libbpf-rs = "=0.26.2"
 log = "0.4"
 nvml-wrapper = "0.12"
 nvml-wrapper-sys = "0.9"
-perf-event-open-sys = "6"
 scx_stats = { path = "../../../rust/scx_stats", version = "1.1.1" }
 scx_stats_derive = { path = "../../../rust/scx_stats/scx_stats_derive", version = "1.1.1" }
 scx_utils = { path = "../../../rust/scx_utils", version = "1.1.1", features = ["autopower", "gpu-topology"] }
